Always check for SOAP faults.
As per Issue #2, sometimes we can get an error back from VIES with a faultcode
and faultstring but a 200 OK HTTP status, which is unhelpful.

So, refactor the checking around - first check for any reported SOAP faults,
then check the HTTP status is OK, and finally expect the response to be
successful and attempt to parse the expected validation results.

TODO: this wants some further refactoring and tidying.  In fact, I might just
start by running the whole file through perltidy :)