Handle some changes in the VIES API
The (broken?) content_type previously set seemed to cause the VIES
server to return a 500 error - and the response is now namespaced.

Updated the generated SOAP envelope based on the request template which
https://www.wsdl-analyzer.com/ generated for me from VIES' .wsdl file.

This module would probably be better re-written to use a SOAP client
rather than hand-rolling XML, but I don't have time to do that, or much
desire to add extra dependencies if I can help it.

This change makes it work again.

Many thanks to Timothy Su and Thomas Glase for reporting the problem.